Create Binary_search.c
In the binary search algorithm, we first sort the array in non-decreasing order. Then, we search for the given key in the middle of the array. If the middle element is equal to the key, we return the index of the middle element. If the middle element is less than the key, then we search for the key in the right half of the array. If the middle element is greater than the key, then we search for the key in the left half of the array. We repeat this process until the key is found or the search space is exhausted.