Create Ternary_search.c
Ternary search is a variation of binary search that works on sorted arrays. Instead of dividing the array in half at each step, it divides the array into three parts. It calculates two midpoints, mid1 and mid2, which divide the array into three equal parts. If x is equal to either arr[mid1] or arr[mid2], we've found the element and return its index. Otherwise, we check whether x is less than arr[mid1] or greater than arr[mid2]. If x is less than arr[mid1], we search the left third of the array. If x is greater than arr[mid2], we search the right third of the array. If x is between arr[mid1] and arr[mid2], we search the middle third of the array. We continue this process recursively until we find x or determine that it's not in the array.