Create level_order_tree.c
Run a for loop for counter i, i.e. current height from 1 to h (height of the tree).
Use DFS to traverse the tree and maintain height for the current node.
If the Node is NULL then return;
If level is 1 print(tree->data);
Else if the level is greater than 1, then
Recursively call to for tree->left, level-1.
Recursively call to for tree->right, level-1.